+++ packages/react-mui-shared/src/components/PangeaDataGrid/components/Search/index.tsx
@@ -11,9 +11,7 @@ import {
   SxProps,
   ButtonProps,
 } from "@mui/material";
-import ConditionalAutocomplete, {
-  ConditionalOption,
-} from "../../../ConditionalAutocomplete";
+import ConditionalAutocomplete from "../../../ConditionalAutocomplete";
 import PopoutCard from "../../../PopoutCard";
 
 import { useInternalState } from "../../../../utils/hooks";
@@ -21,6 +19,8 @@ import FiltersForm, { FilterFormProps } from "./FiltersForm";
 import FiltersBar from "./FiltersBar";
 import ColumnsPopout, { ColumnsPopoutProps } from "./ColumnsPopout";
 import { PDG } from "../../types";
+import { OptionComponentProps } from "../../../ConditionalAutocomplete/AutocompleteOptionComponent";
+import { ConditionalOption } from "../../../ConditionalAutocomplete/types";
 
 export interface SearchProps<FiltersObj> {
   query?: string;
@@ -37,6 +37,9 @@ export interface SearchProps<FiltersObj> {
   StartBarComponent?: ReactNode;
   SearchButtonSx?: SxProps;
   SearchButtonProps?: Partial<ButtonProps>;
+  ConditionalAutocompleteProps?: {
+    OptionComponent?: FC<OptionComponentProps>;
+  };
 }
 
 const Search = <